fix(platform-ui): derive assistant raw responses
