+++ packages/platform-ui/src/pages/AgentsRunScreen.tsx
@@ -17,6 +17,7 @@ import type {
 import { graphSocket } from '@/lib/graph/socket';
 import { notifyError, notifySuccess } from '@/lib/notify';
 import { gatherToolCalls } from '@/lib/toolCalls';
+import { extractLlmResponse } from '@/utils/llmResponse';
 import { formatDuration } from '@/components/agents/runTimelineFormatting';
 
 const EVENT_FILTER_OPTIONS: EventFilter[] = ['message', 'llm', 'tool', 'summary'];
@@ -418,139 +419,7 @@ function extractReasoning(
   return undefined;
 }
 
-function extractTextFromRawResponse(raw: unknown, options?: { ignoreMessage?: boolean }): string | null {
-  const ignoreMessage = options?.ignoreMessage === true;
-  const visited = new WeakSet<object>();
 
-  const extract = (value: unknown): string | null => {
-    if (typeof value === 'string') {
-      const trimmed = value.trim();
-      return trimmed.length > 0 ? value : null;
-    }
-
-    if (Array.isArray(value)) {
-      const parts: string[] = [];
-      for (const item of value) {
-        const text = extract(item);
-        if (typeof text === 'string' && text.length > 0) {
-          parts.push(text);
-        }
-      }
-      if (parts.length > 0) {
-        return parts.join('\n\n');
-      }
-      return null;
-    }
-
-    const record = coerceRecord(value);
-    if (!record) return null;
-    if (visited.has(record)) return null;
-    visited.add(record);
-
-    const directKeys: Array<keyof typeof record> = ['content', 'text', 'output_text', 'outputText'];
-    for (const key of directKeys) {
-      if (key in record) {
-        const text = extract(record[key]);
-        if (typeof text === 'string' && text.length > 0) return text;
-      }
-    }
-
-    if (!ignoreMessage) {
-      if ('message' in record) {
-        const text = extract((record as Record<string, unknown>).message);
-        if (typeof text === 'string' && text.length > 0) return text;
-      }
-
-      if ('messages' in record) {
-        const text = extract((record as Record<string, unknown>).messages);
-        if (typeof text === 'string' && text.length > 0) return text;
-      }
-    }
-
-    const arrayKeys: Array<keyof typeof record> = ['choices', 'outputs', 'output', 'responses'];
-    for (const key of arrayKeys) {
-      if (Array.isArray(record[key])) {
-        for (const entry of record[key] as unknown[]) {
-          const text = extract(entry);
-          if (typeof text === 'string' && text.length > 0) return text;
-        }
-      }
-    }
-
-    if ('delta' in record) {
-      const text = extract((record as Record<string, unknown>).delta);
-      if (typeof text === 'string' && text.length > 0) return text;
-    }
-
-    const nestedKeys: Array<keyof typeof record> = ['data', 'body', 'result', 'response', 'value'];
-    for (const key of nestedKeys) {
-      if (key in record) {
-        const text = extract(record[key]);
-        if (typeof text === 'string' && text.length > 0) return text;
-      }
-    }
-
-    return null;
-  };
-
-  return extract(raw);
-}
-
-function extractLlmResponse(event: RunTimelineEvent): string {
-  if (isNonEmptyString(event.errorMessage)) {
-    return event.errorMessage;
-  }
-
-  const llmCall = event.llmCall;
-  if (!llmCall) return '';
-
-  const responseText = llmCall.responseText;
-  if (isNonEmptyString(responseText)) return responseText;
-
-  const rawResponse = llmCall.rawResponse;
-  if (rawResponse !== null && rawResponse !== undefined) {
-    if (typeof rawResponse === 'string') {
-      const trimmed = rawResponse.trim();
-      if (trimmed.length > 0) return trimmed;
-    }
-
-    const record = coerceRecord(rawResponse);
-    if (record) {
-      const candidateKeys: Array<keyof typeof record> = ['output', 'outputs', 'responses', 'choices', 'result', 'response', 'value'];
-      for (const key of candidateKeys) {
-        if (!(key in record)) continue;
-        const text = extractTextFromRawResponse(record[key], { ignoreMessage: key !== 'choices' });
-        if (isNonEmptyString(text)) return text;
-      }
-    }
-
-    const rawText = extractTextFromRawResponse(rawResponse, { ignoreMessage: true });
-    if (isNonEmptyString(rawText)) return rawText;
-  }
-
-  if (Array.isArray(event.attachments)) {
-    for (const attachment of event.attachments) {
-      if (!attachment || attachment.kind !== 'response') continue;
-
-      const candidates: unknown[] = [];
-      if (attachment.contentText !== undefined && attachment.contentText !== null) {
-        const parsedText = typeof attachment.contentText === 'string' ? parseMaybeJson(attachment.contentText) : attachment.contentText;
-        candidates.push(parsedText);
-      }
-      if (attachment.contentJson !== undefined && attachment.contentJson !== null) {
-        const parsedJson = typeof attachment.contentJson === 'string' ? parseMaybeJson(attachment.contentJson) : attachment.contentJson;
-        candidates.push(parsedJson);
-      }
-
-      for (const candidate of candidates) {
-        const text = extractTextFromRawResponse(candidate, { ignoreMessage: true });
-        if (isNonEmptyString(text)) return text;
-      }
-    }
-  }
-
-  return '';
-}
 
 function normalizeRunEvent(event: RunTimelineEvent, previous?: RunTimelineEvent): RunTimelineEvent {
   const merged: RunTimelineEvent = {
